Refactored each tab t own component
